+++ src/app/containers/Image/helpers/srcSet.js
@@ -1,4 +1,4 @@
-import getIchefUrl from './ichefUrlHelper';
+import getIchefUrl from './ichefUrl';
 
 const createSrcset = (originCode, locator, originalImageWidth) => {
   const resolutions = [240, 320, 480, 624, 800];
